# backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "getOrderDetails",
        "summary": "Get order details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/orders": {
      "get": {
        "operationId": "getAllOrders",
        "summary": "Get all orders (Admin)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "confirmed",
                "processing",
                "ready_to_ship",
                "shipped",
                "out_for_delivery",
                "delivered",
                "completed",
                "cancelled",
                "refunded",
                "returned",
                "payment_failed"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{id}/status": {
      "post": {
        "operationId": "adminSet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "checkout"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "updateStatus",
        "summary": "Update order status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/orders/{id}": {
      "get": {
        "operationId": "getOrderDetails",
        "summary": "Get order details (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{id}/ship": {
      "post": {
        "operationId": "shipOrder",
        "summary": "Ship order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{id}/refund": {
      "post": {
        "operationId": "processRefund",
        "summary": "Process refund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/{id}/notes": {
      "post": {
        "operationId": "addNotes",
        "summary": "Add admin notes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/orders/analytics/summary": {
      "get": {
        "operationId": "getAnalytics",
        "summary": "Get admin analytics",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/carts/analytics": {
      "get": {
        "operationId": "getAnalytics",
        "summary": "Get comprehensive cart analytics",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-carts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/analytics/revenue": {
      "get": {
        "operationId": "getRevenueAnalytics",
        "summary": "Get revenue analytics",
        "description": "Detailed revenue analysis with trends and breakdowns",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "نوع الفترة الزمنية",
            "schema": {
              "enum": [
                "daily",
                "weekly",
                "monthly",
                "quarterly",
                "yearly"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "تاريخ البداية",
            "schema": {
              "example": "2023-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "تاريخ النهاية",
            "schema": {
              "example": "2023-12-31",
              "type": "string"
            }
          },
          {
            "name": "compareWithPrevious",
            "required": false,
            "in": "query",
            "description": "مقارنة مع الفترة السابقة",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue analytics retrieved successfully"
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/analytics/performance": {
      "get": {
        "operationId": "getPerformanceMetrics",
        "summary": "Get system performance metrics",
        "description": "API response times, error rates, uptime, and system health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Performance metrics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceMetricsDto"
                }
              }
            }
          }
        },
        "tags": [
          "analytics"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/exchange-rates/analytics": {
      "get": {
        "operationId": "getAnalytics",
        "summary": "Get exchange rates analytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Analytics retrieved successfully"
          }
        },
        "tags": [
          "admin-exchange-rates"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Tagadod API",
    "description": "Complete e-commerce platform for solar products with advanced analytics and multi-language support",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.tagadod.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "processing",
              "ready_to_ship",
              "shipped",
              "out_for_delivery",
              "delivered",
              "completed",
              "cancelled",
              "refunded",
              "returned",
              "payment_failed"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "ShipOrderDto": {
        "type": "object",
        "properties": {
          "shippingCompany": {
            "type": "string",
            "example": "DHL"
          },
          "trackingNumber": {
            "type": "string",
            "example": "DHL123456789"
          },
          "estimatedDeliveryDate": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "shippingCompany",
          "trackingNumber"
        ]
      },
      "RefundOrderDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 50000
          },
          "reason": {
            "type": "string",
            "example": "عطل في المنتج"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "amount",
          "reason"
        ]
      },
...
      "PerformanceMetricsDto": {
        "type": "object",
        "properties": {
          "apiResponseTime": {
            "type": "number",
            "description": "متوسط وقت استجابة API",
            "example": 245
          },
          "errorRate": {
            "type": "number",
            "description": "معدل الأخطاء",
            "example": 0.02
          },
          "uptime": {
            "type": "number",
            "description": "نسبة التشغيل",
            "example": 99.9
          },
          "concurrentUsers": {
            "type": "number",
            "description": "المستخدمون المتزامنون",
            "example": 1250
          },
          "memoryUsage": {
            "type": "number",
            "description": "ذاكرة النظام المستخدمة",
            "example": 75.5
          },
          "cpuUsage": {
            "type": "number",
            "description": "استخدام المعالج",
            "example": 45.2
          },
          "diskUsage": {
            "type": "number",
            "description": "مساحة القرص المستخدمة",
            "example": 68.3
          },
          "activeConnections": {
            "type": "number",
            "description": "عدد قواعد البيانات النشطة",
            "example": 5
          },
          "slowestEndpoints": {
            "description": "أبطأ نقاط النهاية",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "databaseStats": {
            "type": "object",
            "description": "إحصائيات قاعدة البيانات"
          }
        },
        "required": [
          "apiResponseTime",
          "errorRate",
          "uptime",
          "concurrentUsers",
          "memoryUsage",
          "cpuUsage",
          "diskUsage",
          "activeConnections",
          "slowestEndpoints",
          "databaseStats"
        ]
      },
...
    }
  }
}